Cleaned the project up a lot. Added Shippable and CodeClimate.
