+++ Gemfile
@@ -9,5 +9,7 @@ group :development do
   gem "rspec", "~> 2.8.0"
   gem "rdoc", "~> 3.12"
   gem "bundler", ">= 1.0.0"
-  gem "jeweler", "~> 1.8.3"
+  gem "jeweler", "~> 1.8.8"
 end
+
+gem "codeclimate-test-reporter", group: :test, require: nil
