+++ shippable.yml
@@ -1,0 +1,9 @@
+language: ruby
+rvm:
+  - ruby-1.9.3-head
+  - ruby-head
+  - jruby-head
+script:
+  - CODECLIMATE_REPO_TOKEN=32061319cc6d2e0bd0d0eb88afdc6d7b68829b357ca76efb99f86d7a1a89539e bundle exec rspec
+notifications:
+  email: false
